requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.2
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
Django==2.2.6
django-allauth==0.40.0
django-bootstrap4==1.0.1
django-extensions==2.2.5
django-test-plus==1.3.1
djangorestframework==3.10.3
Faker==2.0.3
future==0.18.2
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.25.0
h5py==2.10.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
parso==0.5.1
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
pylint==2.4.3
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-datasets==1.3.0
tensorflow-estimator==2.0.1
tensorflow-metadata==0.15.1
termcolor==1.1.0
text-unidecode==1.3
tqdm==4.39.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
